mlx5: update timecounter at least twice per counter overflow
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Miroslav Lichvar <mlichvar@redhat.com>
commit 5d8678365c90b9ce1fd2243ff5ea562609f6cec1

The timecounter needs to be updated at least once in half of the
cyclecounter interval to prevent timecounter_cyc2time() interpreting a
new timestamp as an old value and causing a backward jump.

This would be an issue if the timecounter multiplier was so small that
the update interval would not be limited by the 64-bit overflow in
multiplication.

Shorten the calculated interval to make sure the timecounter is updated
in time even when the system clock is slowed down by up to 10%, the
multiplier is increased by up to 10%, and the scheduled overflow check
is late by 15%.

	Cc: Richard Cochran <richardcochran@gmail.com>
	Cc: Ariel Levkovich <lariel@mellanox.com>
	Cc: Saeed Mahameed <saeedm@mellanox.com>
	Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
	Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>
(cherry picked from commit 5d8678365c90b9ce1fd2243ff5ea562609f6cec1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>