IB/mlx5: Fix scatter to CQE in DCT QP creation
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Guy Levi <guyle@mellanox.com>
commit 7249c8ea227a582c14f63e9e8853eb7369122f10

When scatter to CQE is enabled on a DCT QP it corrupts the mailbox command
since it tried to treat it as as QP create mailbox command instead of a
DCT create command.

The corrupted mailbox command causes userspace to malfunction as the
device doesn't create the QP as expected.

A new mlx5 capability is exposed to user-space which ensures that it will
not enable the feature on DCT without this fix in the kernel.

Fixes: 5d6ff1babe78 ("IB/mlx5: Support scatter to CQE for DC transport type")
	Signed-off-by: Guy Levi <guyle@mellanox.com>
	Signed-off-by: Leon Romanovsky <leonro@mellanox.com>
	Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
(cherry picked from commit 7249c8ea227a582c14f63e9e8853eb7369122f10)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>