ipv6: Fix dump of specific table with strict checking
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author David Ahern <dsahern@gmail.com>
commit 73155879b3c1ac3ace35208a54a3a160ec520bef

Dump of a specific table with strict checking enabled is looping. The
problem is that the end of the table dump is not marked in the cb. When
dumping a specific table, cb args 0 and 1 are not used (they are the hash
index and entry with an hash table index when dumping all tables). Re-use
args[0] to hold a 'done' flag for the specific table dump.

Fixes: 13e38901d46ca ("net/ipv6: Plumb support for filtering route dumps")
	Reported-by: Jakub Kicinski <jakub.kicinski@netronome.com>
	Signed-off-by: David Ahern <dsahern@gmail.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 73155879b3c1ac3ace35208a54a3a160ec520bef)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>