rhashtable: simplify nested_table_alloc() and rht_bucket_nested_insert()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author NeilBrown <neilb@suse.com>
commit 5af68ef7333c8606bfe6e400cb962081518c3acb

Now that we don't use the hash value or shift in nested_table_alloc()
there is room for simplification.
We only need to pass a "is this a leaf" flag to nested_table_alloc(),
and don't need to track as much information in
rht_bucket_nested_insert().

Note there is another minor cleanup in nested_table_alloc() here.
The number of elements in a page of "union nested_tables" is most naturally

  PAGE_SIZE / sizeof(ntbl[0])

The previous code had

  PAGE_SIZE / sizeof(ntbl[0].bucket)

which happens to be the correct value only because the bucket uses all
the space in the union.

	Acked-by: Herbert Xu <herbert@gondor.apana.org.au>
	Signed-off-by: NeilBrown <neilb@suse.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 5af68ef7333c8606bfe6e400cb962081518c3acb)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>