ipv6: Fix PMTU updates for UDP/raw sockets in presence of VRF
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author David Ahern <dsahern@gmail.com>
commit 7ddacfa564870cdd97275fd87decb6174abc6380

Preethi reported that PMTU discovery for UDP/raw applications is not
working in the presence of VRF when the socket is not bound to a device.
The problem is that ip6_sk_update_pmtu does not consider the L3 domain
of the skb device if the socket is not bound. Update the function to
set oif to the L3 master device if relevant.

Fixes: ca254490c8df ("net: Add VRF support to IPv6 stack")
	Reported-by: Preethi Ramachandra <preethir@juniper.net>
	Signed-off-by: David Ahern <dsahern@gmail.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7ddacfa564870cdd97275fd87decb6174abc6380)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>