ipv6: Consider sk_bound_dev_if when binding a raw socket to an address
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Mike Manning <mmanning@vyatta.att-mail.com>
commit 72f7cfab6f93a8ea825fab8ccfb016d064269f7f

IPv6 does not consider if the socket is bound to a device when binding
to an address. The result is that a socket can be bound to eth0 and
then bound to the address of eth1. If the device is a VRF, the result
is that a socket can only be bound to an address in the default VRF.

Resolve by considering the device if sk_bound_dev_if is set.

	Signed-off-by: Mike Manning <mmanning@vyatta.att-mail.com>
	Reviewed-by: David Ahern <dsahern@gmail.com>
	Tested-by: David Ahern <dsahern@gmail.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 72f7cfab6f93a8ea825fab8ccfb016d064269f7f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>