nvme-tcp: support C2HData with SUCCESS flag
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Sagi Grimberg <sagi@grimberg.me>
commit 602d674ce90f64ac135452fb9b2b058acb53b226

A C2HData PDU with the SUCCESS flag set indicates that the I/O was
completed by the controller successfully and means that a subsequent
completion response capsule PDU will be ommitted.

If we see this flag, fisrt we check that LAST_PDU flag is set as well,
and then we complete the request when the data transfer (and data digest
verification if its on) is done.

While we're at it, reuse a bit of code with nvme_fail_request.

	Reported-by: Steve Blightman <steve.blightman@oracle.com>
	Suggested-by: Oliver Smith-Denny <osmithde@cisco.com>
	Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
	Reviewed-by: Oliver Smith-Denny <osmithde@cisco.com>
	Tested-by: Oliver Smith-Denny <osmithde@cisco.com>
	Signed-off-by: Christoph Hellwig <hch@lst.de>
	Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit 602d674ce90f64ac135452fb9b2b058acb53b226)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>