mlxsw: spectrum_switchdev: Ban PVID change if bridge has a RIF
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Petr Machata <petrm@mellanox.com>
commit 567ad1a2bb6c1e2ca9644d1b47263b7d5e1ff900

When traffic passes through a router port, it needs to be assigned a FID
for ASIC to forward correctly. For bridges, this FID used to be the one
corresponding to VLAN 1. In a previous patch, this was changed to
instead use the PVID at the time that the RIF is created. This patch
guards PVID changes after the RIF was introduced.

	Signed-off-by: Petr Machata <petrm@mellanox.com>
	Signed-off-by: Ido Schimmel <idosch@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 567ad1a2bb6c1e2ca9644d1b47263b7d5e1ff900)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>