selftests: mlxsw: Add tc flower scale test
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Petr Machata <petrm@mellanox.com>
commit d67a94e81fda793fc2cd8562501117720e221634

Add test of capacity to offload flower.

This is a generic portion of the test that is meant to be called from a
driver that supplies a particular number of rules to be tested with.

	Signed-off-by: Petr Machata <petrm@mellanox.com>
	Reviewed-by: Yuval Mintz <yuvalm@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d67a94e81fda793fc2cd8562501117720e221634)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>