mlxsw: reg: Add Policy-Engine Region eRP Register
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Ido Schimmel <idosch@mellanox.com>
commit f1c7d9cce2a8ec3d6223a2d7b16c6a9c00c52deb

The PERERP register configures the region eRPs. It can be used, for
example, to enable lookup in the C-TCAM in addition to the A-TCAM.

To be able to perform a lookup in the C-TCAM we need to "use" the eRP
table. This is done by marking the pointer as valid, but zeroing the eRP
table vector.

	Signed-off-by: Jiri Pirko <jiri@mellanox.com>
	Signed-off-by: Ido Schimmel <idosch@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f1c7d9cce2a8ec3d6223a2d7b16c6a9c00c52deb)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>