mlxsw: spectrum_switchdev: Prepare function for VLAN-aware bridges
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Ido Schimmel <idosch@mellanox.com>
commit 48fde4660654d5bbaa09c2a38fa9ca7c72f1a95e

The vxlan_join() function resolves the FID on which the VNI should be
set and then sets the VNI. Currently, the FID is simply resolved
according to the ifindex of the bridge device to which the VxLAN device
is enslaved. This works because only VLAN-unaware bridges are supported.

With VLAN-aware bridges the FID would need to be resolved based on the
VLAN to which the VNI is mapped to.

Add the VLAN ID to the argument list of the function.

	Signed-off-by: Ido Schimmel <idosch@mellanox.com>
	Reviewed-by: Petr Machata <petrm@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 48fde4660654d5bbaa09c2a38fa9ca7c72f1a95e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>