selftests: forwarding: Add PVID test case for VXLAN with VLAN-aware bridges
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-147.el8
commit-author Ido Schimmel <idosch@mellanox.com>
commit 55939b262a61d35d3ccde8a388976b946d654dcf

When using VLAN-aware bridges with VXLAN, the VLAN that is mapped to the
VNI of the VXLAN device is that which is configured as "pvid untagged"
on the corresponding bridge port.

When these flags are toggled or when the VLAN is deleted entirely,
remote hosts should not be able to receive packets from the VTEP.

Add a test case for above mentioned scenarios.

	Signed-off-by: Petr Machata <petrm@mellanox.com>
	Signed-off-by: Ido Schimmel <idosch@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 55939b262a61d35d3ccde8a388976b946d654dcf)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>